Add default page
