+++ lib/screens/home.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:open_erp/components/nothing_to_show.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Home Page'),
+      ),
+      body: Column(
+        mainAxisSize: MainAxisSize.max,
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: const [
+          NothingToShow(),
+        ],
+      ),
+    );
+  }
+}
